% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_nvc.R
\name{assign_nvc}
\alias{assign_nvc}
\title{Determine most similar NVCs for each sample by comparing samples against NVC psuedo-quadrat data}
\usage{
assign_nvc(
  samp_df,
  comp_df = NULL,
  spp_col = "species",
  samp_id = "ID",
  comp_id = "Pid3",
  top_n = 20
)
}
\arguments{
\item{samp_df}{A \code{data.frame} containing the sample data that you want to assign to NVC via comparison with psuedo-quadrat data}

\item{comp_df}{A \code{data.frame} containing the comparison data samples you want to match your focal samples in \code{samp_df} to. The default value of \code{NULL} means that the comparison will use the NVC pseudo-quadrat data included with the package/}

\item{spp_col}{A character vector specifying the name of the column in \code{samp_df} and \code{comp_df}that contain the species names or codes that you want to match on. The default is for the column to be named \code{species}.}

\item{samp_id}{A charcter vector specifying the name of the column in \code{samp_df} that contains the IDs identifying the data for each sample in the data you want assigned to NVC. The default is for the column to be named \code{ID}.}

\item{comp_id}{A charcter vector specifying the name of the column in \code{comp_df} that contains the IDs identifying the data for each sample in the comparison data. The default is for the column to be named \code{ID}.}

\item{top_n}{An integer specifying the number most similar comparison samples to be returned, or \code{NULL} to return all matches. The default \code{top_n = 20} will return the 20 most similar samples from \code{comp_df} for each sample in \code{samp_df}.}
}
\value{
a \code{data.frame} detailing the most similar NVC psuedo-quadrats for each sample in the focal data \code{samp_df}
}
\description{
Determine most similar NVCs for each sample by comparing samples against NVC psuedo-quadrat data
}
\examples{

# Create a test dataset
test_samp = data.frame(ID = 1, species = c("Alchemilla alpina","Anthoxanthum odoratum","Blechnum spicant","Deschampsia flexuosa","Galium saxatile","Luzula sylvatica","Melampyrum pratense","Nardus stricta","Potentilla erecta","Rumex acetosella","Vaccinium myrtillus","Dicranum majus","Hylocomium splendens","Isothecium myosuroides","Plagiothecium undulatum","Pleurozium schreberi","Polytrichum commune","Racomitrium lanuginosum","Rhytidiadelphus loreus","Calypogeia fissa","Cladonia arbuscula","Cladonia bellidiflora","Cladonia coccifera","Cladonia uncialis"))

# Now run assign_nvc on this data to get the top 10 most similar
test_jac = assign_nvc(test_samp, top_n=10)

}
