% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nvc_pquads}
\alias{nvc_pquads}
\title{Psuedo-quadrat data for GB NVC Comunities (updated)}
\format{
A data frame with 974601 rows and 6 variables:
\describe{
\item{name}{species scientific name}
\item{spp}{species code}
\item{Pid2}{psuedo-quadrat ID}
\item{Pid3}{ID created concatenating NVC and psuedo-quadrat ID \code{Pid2}}
\item{Pid3}{psued-quadrat number}
}
}
\source{
\url{http://dx.doi.org/10.1016/j.envpol.2013.04.008}
}
\usage{
nvc_pquads
}
\description{
A dataset containing pseudo-quadrat data for NVC
communities generated by random sampling from NVC species
list. The psuedo-quadrat data is based on that from a study by Tipping \emph{et al.} 2013.
but has been updated to include additional NVCs
}
\details{
When generating psudeo-quadrants the  probabilities
of species being in a sample were set to match the
constancy scores of the species in that given NVC
community/sub-community. The probabilities associated
with each constancy score were 0.1, 0.3, 0.5, 0.7 and 0.9 for
constancy scores 1, 2, 3, 4 and 5 respectively. The number
of psuedo-quadrats generated for each NVC community or
sub-community was set to match the number of samples originally
used to define the NVC and produce the contingency tables.

The citation for the Tipping \emph{et al.} 2013 paper is:
Tipping, E., Henrys, P. A., Maskell, L. C., & Smart, S. M. (2013). Nitrogen deposition effects on plant species diversity; Threshold loads from field data. \emph{Environmental Pollution}, 179, 218-223.
}
\keyword{datasets}
