% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nvc_samp_sizes}
\alias{nvc_samp_sizes}
\title{NVC Sample sizes}
\format{
A data frame with 879 rows and 8 variables:
\describe{
\item{NVC}{The NVC community or sub-community code}
\item{LEVEL}{The level of the NVC given, 1 for community, 2 for sub-community}
\item{NVC_COMMUNITY}{The NVC community that the NVC belongs to}
\item{VEG_TYPE}{The vegetation type letter code for the NVC, i.e. the first letter(s)}
\item{NVC_NUMBER}{The number of that NVC within the current \code{VEG_TYPE}, i.e. the digit(s) after the VEG_TYPE in the NVC code}
\item{NVC_SUB_CODE}{The community code for that NVC, if applicable}
\item{VERSION}{The version of that NVC as some NVC have been updated, with new sub-communities being added meaning that for some communities there is an original version and then an updated v2 version that include additional species/samples from the added subcategories}
\item{N_SAMPLES}{The number of samples original used in the data used to define the NVC community/sub-community}
}
}
\usage{
nvc_samp_sizes
}
\description{
A dataset listing the sizes of the samples used to define that NVC.
The psuedo-quadrat included with the package has been generated so that the number of
psuedo-quadrats for each NVC matches the original data sample sizes.
}
\details{
Using the data in dataset a new set(s) of psuedo-quadrats that match the original sample
sizes can be generated
}
\keyword{datasets}
