% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psuedo_quad.R
\name{psuedo_quad}
\alias{psuedo_quad}
\title{Generate psuedo-quadrat data from a data.frame of species with occurrence frequencies}
\usage{
psuedo_quad(
  sp_freq,
  nq,
  spp_col = "species",
  freq_col = "freq",
  nvc_col = NULL
)
}
\arguments{
\item{sp_freq}{A 2 column \code{data.frame} containing the species names, or codes, along with the frequency probability of occurrence}

\item{nq}{An integer specifying the number of psuedo-quadrats to create. Where \code{sp_freq} contains data for multiple NVCs/groupings, indicated by values in \code{nvc_col} a vector of integer values, can be supplied if you want different numbers of psuedo-quadrats for the different NVC/groupings.  In this case the values in the \code{nq} vector must match the order in which the NVCs/groupings in \code{nvc_col} would be sorted, i.e. the 1st nq value is the number of quadrats for the 1st NVC/group in \code{nvc_col}, the 2nd nq value the number of quadrats for the 2nd NVC, etc.}

\item{spp_col}{A character string giving the name of the column in sp_freq that contains the species names/codes. Default is "species"}

\item{freq_col}{A character string giving the name of the column in sp_freq that contains the probability of occurrence for the give species. Default is "freq"}

\item{nvc_col}{A character string giving the name of the column in \code{sp_freq} that contains the NVC codes (or other groupings). When specified
it assumes that \code{sp_freq} contains separate species lists and associated frequencies for a number of different NVC communities or other groupings with that column specifying
the NVC/group to which each row belongs . Where this is not null a separate set of \code{nq} number of psudeo-quadrats will be generated
for each NVC/group in \code{nvc_col} of \code{sp_freq}}
}
\value{
A 2 column \code{data.frame} with the first column giving the pseudo-quadrat number and 2nd the species names/codes of species present in the psuedo-quadrat
}
\description{
Produce data for psuedo-quadrat(s) from a species via random sampling using occurrence frequencies/probabilities of occurrence for each species to produce random samples of species that could have been obtained
}
\examples{
# Setup a data.frame with a species list and occurrency frequencies

  sp_freq <- data.frame(species = paste("Species",1:10),freq = c(0.9,0.8,0.4,0.4,0.4,0.3,0.1,0.1,0.1,0.1))

# Now produce 5 pseudo-quadrat samples based on this composition data.frame

  test_ps <- psuedo_quad(sp_freq,5)


# Now an example where the data.frame contains species lists & frequencies for multiple NVCs each of which you want to generate psuedo-quarats for
 # Load NVC_communites dataset
   data(NVC_communities)
 # Subset to the first 5 NVCs
   sp_freq = subset(NVC_communities,NVC \%in\% unique(NVC_communities$NVC)[1:5])
 # Create psuedo-quarats for these NVCs (note spp_col in NVC communities is Species not species as expected by default)
   test_ps = psuedo_quad(sp_freq,nq = c(10,5,5,5,5),nvc_col = "NVC",spp_col = "Species")

}
