% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_sim.R
\name{jaccard_sim}
\alias{jaccard_sim}
\title{Calculate Jaccard similarity between a focal sample and one or more comparison sample(s)}
\usage{
jaccard_sim(focal_spp, comp_spp_list, focal_id = NULL, top_n = 20)
}
\arguments{
\item{focal_spp}{A character vector containing species for the focal sample that to be compared against against the species in \code{comp_spp_list}. The species can be specified using species names or ID codes but need to correspond to names or codes used in \code{comp_spp_list}.}

\item{comp_spp_list}{A \code{list} containing the species recorded in the comparison samples, where each element is a different comparison sample and contains a character vector with the species recorded for that comparison sample. The names of the list elements will be the names/identifiers for the comparison samples. As with \code{focal_spp} the species can be specified using either names or ID codes.}

\item{focal_id}{An name or identifier for the focal sample. Default is \code{NULL}}

\item{top_n}{An integer determining how many of the top scoring comparison samples and their similarity scores will be returned in the final data.frame. The default is \code{50}}
}
\value{
A \code{data.frame} containing the comparison samples, similarity scores for the \code{top_n} matches from \code{comp_spp_list}
}
\description{
Calculate the Jaccard similarity values between the species recorded for a focal site or sample against those recorded for a series of comparison sites or samples (e.g. against the NVC pseudo-quadrats) and then return the \code{top_n} matches in \code{comp_spp_list} based on their similarity scores with the species in focal sample
}
