% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psuedo_quad.R
\name{psuedo_quad}
\alias{psuedo_quad}
\title{Generate psuedo-quadrat data from a data.frame of species with occurrence frequencies}
\usage{
psuedo_quad(nq, sp_freq, spp_col = "species", freq_col = "freq")
}
\arguments{
\item{nq}{The number of psuedo-quadrats to create}

\item{sp_freq}{A 2 column \code{data.frame} containing the species names, or codes, along with the frequency probability of occurrence}

\item{spp_col}{A character string giving the name of the column in sp_freq that contains the species names/codes. Default is "species"}

\item{freq_col}{A character string giving the name of the column in sp_freq that contains the probility of occurrence for the give species. Default is "freq"}
}
\value{
A 2 column \code{data.frame} with the first column giving the pseudo-quadrat number and 2nd the species names/codes of species present in the psuedo-quadrat
}
\description{
Produce data for psuedo-quadrat(s) from a species via random sampling using occurrence frequencies/probabilities of occurrence for each species to produce random samples of species that could have been obtained
}
\examples{
# Setup a data.frame with a species list and occurrency frequencies

  sp_freq <- data.frame(species = paste("Species",1:10),freq = c(0.9,0.8,0.4,0.4,0.4,0.3,0.1,0.1,0.1,0.1))

# Now produce 5 pseudo-quadrat samples based on this composition data.frame

  test_ps <- psuedo_quad(5,sp_freq)


}
